+++ opentelemetry-sdk/tests/trace/test_trace.py
@@ -17,6 +17,7 @@ from unittest import mock
 
 from opentelemetry import trace as trace_api
 from opentelemetry.sdk import trace
+from opentelemetry.trace import sampling
 from opentelemetry.util import time_ns
 
 
@@ -26,23 +27,60 @@ class TestTracer(unittest.TestCase):
         self.assertIsInstance(tracer, trace_api.Tracer)
 
 
+class TestTracerSampling(unittest.TestCase):
+    def test_default_sampler(self):
+        tracer = trace.Tracer()
+
+        # Check that the default tracer creates real spans via the default
+        # sampler
+        root_span = tracer.create_span(name="root span", parent=None)
+        self.assertIsInstance(root_span, trace.Span)
+        child_span = tracer.create_span(name="child span", parent=root_span)
+        self.assertIsInstance(child_span, trace.Span)
+
+    def test_sampler_no_sampling(self):
+        tracer = trace.Tracer()
+        tracer.sampler = sampling.ALWAYS_OFF
+
+        # Check that the default tracer creates no-op spans if the sampler
+        # decides not to sampler
+        root_span = tracer.create_span(name="root span", parent=None)
+        self.assertIsInstance(root_span, trace_api.DefaultSpan)
+        child_span = tracer.create_span(name="child span", parent=root_span)
+        self.assertIsInstance(child_span, trace_api.DefaultSpan)
+
+
 class TestSpanCreation(unittest.TestCase):
+    def test_create_span_invalid_spancontext(self):
+        """If an invalid span context is passed as the parent, the created
+        span should use a new span id.
+
+        Invalid span contexts should also not be added as a parent. This
+        eliminates redundant error handling logic in exporters.
+        """
+        tracer = trace.Tracer("test_create_span_invalid_spancontext")
+        new_span = tracer.create_span(
+            "root", parent=trace_api.INVALID_SPAN_CONTEXT
+        )
+        self.assertTrue(new_span.context.is_valid())
+        self.assertIsNone(new_span.parent)
+
     def test_start_span_implicit(self):
         tracer = trace.Tracer("test_start_span_implicit")
 
         self.assertIsNone(tracer.get_current_span())
 
-        with tracer.start_span("root") as root:
-            self.assertIs(tracer.get_current_span(), root)
+        root = tracer.start_span("root")
+        self.assertIsNotNone(root.start_time)
+        self.assertIsNone(root.end_time)
+        self.assertEqual(root.kind, trace_api.SpanKind.INTERNAL)
 
-            self.assertIsNotNone(root.start_time)
-            self.assertIsNone(root.end_time)
-            self.assertEqual(root.kind, trace_api.SpanKind.INTERNAL)
+        with tracer.use_span(root, True):
+            self.assertIs(tracer.get_current_span(), root)
 
             with tracer.start_span(
                 "child", kind=trace_api.SpanKind.CLIENT
             ) as child:
-                self.assertIs(tracer.get_current_span(), child)
                 self.assertIs(child.parent, root)
                 self.assertEqual(child.kind, trace_api.SpanKind.CLIENT)
 
@@ -64,9 +102,9 @@ class TestSpanCreation(unittest.TestCase):
                     root_context.trace_options, child_context.trace_options
                 )
 
-            # After exiting the child's scope the parent should become the
-            # current span again.
-            self.assertIs(tracer.get_current_span(), root)
+                # Verify start_span() did not set the current span.
+                self.assertIs(tracer.get_current_span(), root)
+
             self.assertIsNotNone(child.end_time)
 
         self.assertIsNone(tracer.get_current_span())
@@ -82,26 +120,25 @@ class TestSpanCreation(unittest.TestCase):
 
         self.assertIsNone(tracer.get_current_span())
 
+        root = tracer.start_span("root")
+        self.assertIsNotNone(root.start_time)
+        self.assertIsNone(root.end_time)
+
         # Test with the implicit root span
-        with tracer.start_span("root") as root:
+        with tracer.use_span(root, True):
             self.assertIs(tracer.get_current_span(), root)
 
-            self.assertIsNotNone(root.start_time)
-            self.assertIsNone(root.end_time)
-
             with tracer.start_span("stepchild", other_parent) as child:
-                # The child should become the current span as usual, but its
-                # parent should be the one passed in, not the
-                # previously-current span.
-                self.assertIs(tracer.get_current_span(), child)
+                # The child's parent should be the one passed in,
+                # not the current span.
                 self.assertNotEqual(child.parent, root)
                 self.assertIs(child.parent, other_parent)
 
                 self.assertIsNotNone(child.start_time)
                 self.assertIsNone(child.end_time)
 
-                # The child should inherit its context fromr the explicit
-                # parent, not the previously-current span.
+                # The child should inherit its context from the explicit
+                # parent, not the current span.
                 child_context = child.get_context()
                 self.assertEqual(other_parent.trace_id, child_context.trace_id)
                 self.assertNotEqual(
@@ -114,6 +151,60 @@ class TestSpanCreation(unittest.TestCase):
                     other_parent.trace_options, child_context.trace_options
                 )
 
+                # Verify start_span() did not set the current span.
+                self.assertIs(tracer.get_current_span(), root)
+
+            # Verify ending the child did not set the current span.
+            self.assertIs(tracer.get_current_span(), root)
+            self.assertIsNotNone(child.end_time)
+
+    def test_start_as_current_span_implicit(self):
+        tracer = trace.Tracer("test_start_as_current_span_implicit")
+
+        self.assertIsNone(tracer.get_current_span())
+
+        with tracer.start_as_current_span("root") as root:
+            self.assertIs(tracer.get_current_span(), root)
+
+            with tracer.start_as_current_span("child") as child:
+                self.assertIs(tracer.get_current_span(), child)
+                self.assertIs(child.parent, root)
+
+            # After exiting the child's scope the parent should become the
+            # current span again.
+            self.assertIs(tracer.get_current_span(), root)
+            self.assertIsNotNone(child.end_time)
+
+        self.assertIsNone(tracer.get_current_span())
+        self.assertIsNotNone(root.end_time)
+
+    def test_start_as_current_span_explicit(self):
+        tracer = trace.Tracer("test_start_as_current_span_explicit")
+
+        other_parent = trace_api.SpanContext(
+            trace_id=0x000000000000000000000000DEADBEEF,
+            span_id=0x00000000DEADBEF0,
+        )
+
+        self.assertIsNone(tracer.get_current_span())
+
+        # Test with the implicit root span
+        with tracer.start_as_current_span("root") as root:
+            self.assertIs(tracer.get_current_span(), root)
+
+            self.assertIsNotNone(root.start_time)
+            self.assertIsNone(root.end_time)
+
+            with tracer.start_as_current_span(
+                "stepchild", other_parent
+            ) as child:
+                # The child should become the current span as usual, but its
+                # parent should be the one passed in, not the
+                # previously-current span.
+                self.assertIs(tracer.get_current_span(), child)
+                self.assertNotEqual(child.parent, root)
+                self.assertIs(child.parent, other_parent)
+
             # After exiting the child's scope the last span on the stack should
             # become current, not the child's parent.
             self.assertNotEqual(tracer.get_current_span(), other_parent)
@@ -122,30 +213,15 @@ class TestSpanCreation(unittest.TestCase):
 
 
 class TestSpan(unittest.TestCase):
+    def setUp(self):
+        self.tracer = trace.Tracer("test_span")
+
     def test_basic_span(self):
         span = trace.Span("name", mock.Mock(spec=trace_api.SpanContext))
         self.assertEqual(span.name, "name")
 
-    def test_span_members(self):
-        tracer = trace.Tracer("test_span_members")
-
-        other_context1 = trace_api.SpanContext(
-            trace_id=trace.generate_trace_id(),
-            span_id=trace.generate_span_id(),
-        )
-        other_context2 = trace_api.SpanContext(
-            trace_id=trace.generate_trace_id(),
-            span_id=trace.generate_span_id(),
-        )
-        other_context3 = trace_api.SpanContext(
-            trace_id=trace.generate_trace_id(),
-            span_id=trace.generate_span_id(),
-        )
-
-        self.assertIsNone(tracer.get_current_span())
-
-        with tracer.start_span("root") as root:
-            # attributes
+    def test_attributes(self):
+        with self.tracer.start_as_current_span("root") as root:
             root.set_attribute("component", "http")
             root.set_attribute("http.method", "GET")
             root.set_attribute(
@@ -172,27 +248,57 @@ class TestSpan(unittest.TestCase):
             self.assertEqual(root.attributes["misc.pi"], 3.14)
             self.assertEqual(root.attributes["attr-key"], "attr-value2")
 
-            # events
+    def test_events(self):
+        self.assertIsNone(self.tracer.get_current_span())
+
+        with self.tracer.start_as_current_span("root") as root:
+            # only event name
             root.add_event("event0")
-            root.add_event("event1", {"name": "birthday"})
+
+            # event name and attributes
             now = time_ns()
+            root.add_event("event1", {"name": "pluto"})
+
+            # event name, attributes and timestamp
+            now = time_ns()
+            root.add_event("event2", {"name": "birthday"}, now)
+
+            # lazy event
             root.add_lazy_event(
-                trace_api.Event("event2", now, {"name": "hello"})
+                trace_api.Event("event3", {"name": "hello"}, now)
             )
 
-            self.assertEqual(len(root.events), 3)
+            self.assertEqual(len(root.events), 4)
 
             self.assertEqual(root.events[0].name, "event0")
             self.assertEqual(root.events[0].attributes, {})
 
             self.assertEqual(root.events[1].name, "event1")
-            self.assertEqual(root.events[1].attributes, {"name": "birthday"})
+            self.assertEqual(root.events[1].attributes, {"name": "pluto"})
 
             self.assertEqual(root.events[2].name, "event2")
-            self.assertEqual(root.events[2].attributes, {"name": "hello"})
+            self.assertEqual(root.events[2].attributes, {"name": "birthday"})
             self.assertEqual(root.events[2].timestamp, now)
 
-            # links
+            self.assertEqual(root.events[3].name, "event3")
+            self.assertEqual(root.events[3].attributes, {"name": "hello"})
+            self.assertEqual(root.events[3].timestamp, now)
+
+    def test_links(self):
+        other_context1 = trace_api.SpanContext(
+            trace_id=trace.generate_trace_id(),
+            span_id=trace.generate_span_id(),
+        )
+        other_context2 = trace_api.SpanContext(
+            trace_id=trace.generate_trace_id(),
+            span_id=trace.generate_span_id(),
+        )
+        other_context3 = trace_api.SpanContext(
+            trace_id=trace.generate_trace_id(),
+            span_id=trace.generate_span_id(),
+        )
+
+        with self.tracer.start_as_current_span("root") as root:
             root.add_link(other_context1)
             root.add_link(other_context2, {"name": "neighbor"})
             root.add_lazy_link(
@@ -219,6 +325,8 @@ class TestSpan(unittest.TestCase):
             )
             self.assertEqual(root.links[2].attributes, {"component": "http"})
 
+    def test_update_name(self):
+        with self.tracer.start_as_current_span("root") as root:
             # name
             root.update_name("toor")
             self.assertEqual(root.name, "toor")
@@ -235,6 +343,24 @@ class TestSpan(unittest.TestCase):
         span.start()
         self.assertEqual(start_time, span.start_time)
 
+        # default status
+        self.assertTrue(span.status.is_ok)
+        self.assertIs(
+            span.status.canonical_code, trace_api.status.StatusCanonicalCode.OK
+        )
+        self.assertIs(span.status.description, None)
+
+        # status
+        new_status = trace_api.status.Status(
+            trace_api.status.StatusCanonicalCode.CANCELLED, "Test description"
+        )
+        span.set_status(new_status)
+        self.assertIs(
+            span.status.canonical_code,
+            trace_api.status.StatusCanonicalCode.CANCELLED,
+        )
+        self.assertIs(span.status.description, "Test description")
+
     def test_span_override_start_and_end_time(self):
         """Span sending custom start_time and end_time values"""
         span = trace.Span("name", mock.Mock(spec=trace_api.SpanContext))
@@ -247,14 +373,13 @@ class TestSpan(unittest.TestCase):
 
     def test_ended_span(self):
         """"Events, attributes are not allowed after span is ended"""
-        tracer = trace.Tracer("test_ended_span")
 
         other_context1 = trace_api.SpanContext(
             trace_id=trace.generate_trace_id(),
             span_id=trace.generate_span_id(),
         )
 
-        with tracer.start_span("root") as root:
+        with self.tracer.start_as_current_span("root") as root:
             # everything should be empty at the beginning
             self.assertEqual(len(root.attributes), 0)
             self.assertEqual(len(root.events), 0)
@@ -280,6 +405,19 @@ class TestSpan(unittest.TestCase):
 
             root.update_name("xxx")
             self.assertEqual(root.name, "root")
+
+            new_status = trace_api.status.Status(
+                trace_api.status.StatusCanonicalCode.CANCELLED,
+                "Test description",
+            )
+            root.set_status(new_status)
+            # default status
+            self.assertTrue(root.status.is_ok)
+            self.assertEqual(
+                root.status.canonical_code,
+                trace_api.status.StatusCanonicalCode.OK,
+            )
+            self.assertIs(root.status.description, None)
 
 
 def span_event_start_fmt(span_processor_name, span_name):
@@ -313,9 +451,9 @@ class TestSpanProcessor(unittest.TestCase):
         sp1 = MySpanProcessor("SP1", spans_calls_list)
         sp2 = MySpanProcessor("SP2", spans_calls_list)
 
-        with tracer.start_span("foo"):
-            with tracer.start_span("bar"):
-                with tracer.start_span("baz"):
+        with tracer.start_as_current_span("foo"):
+            with tracer.start_as_current_span("bar"):
+                with tracer.start_as_current_span("baz"):
                     pass
 
         # at this point lists must be empty
@@ -324,13 +462,13 @@ class TestSpanProcessor(unittest.TestCase):
         # add single span processor
         tracer.add_span_processor(sp1)
 
-        with tracer.start_span("foo"):
+        with tracer.start_as_current_span("foo"):
             expected_list.append(span_event_start_fmt("SP1", "foo"))
 
-            with tracer.start_span("bar"):
+            with tracer.start_as_current_span("bar"):
                 expected_list.append(span_event_start_fmt("SP1", "bar"))
 
-                with tracer.start_span("baz"):
+                with tracer.start_as_current_span("baz"):
                     expected_list.append(span_event_start_fmt("SP1", "baz"))
 
                 expected_list.append(span_event_end_fmt("SP1", "baz"))
@@ -347,15 +485,15 @@ class TestSpanProcessor(unittest.TestCase):
         # go for multiple span processors
         tracer.add_span_processor(sp2)
 
-        with tracer.start_span("foo"):
+        with tracer.start_as_current_span("foo"):
             expected_list.append(span_event_start_fmt("SP1", "foo"))
             expected_list.append(span_event_start_fmt("SP2", "foo"))
 
-            with tracer.start_span("bar"):
+            with tracer.start_as_current_span("bar"):
                 expected_list.append(span_event_start_fmt("SP1", "bar"))
                 expected_list.append(span_event_start_fmt("SP2", "bar"))
 
-                with tracer.start_span("baz"):
+                with tracer.start_as_current_span("baz"):
                     expected_list.append(span_event_start_fmt("SP1", "baz"))
                     expected_list.append(span_event_start_fmt("SP2", "baz"))
 
@@ -380,9 +518,9 @@ class TestSpanProcessor(unittest.TestCase):
         # Span processors are created but not added to the tracer yet
         sp = MySpanProcessor("SP1", spans_calls_list)
 
-        with tracer.start_span("foo"):
-            with tracer.start_span("bar"):
-                with tracer.start_span("baz"):
+        with tracer.start_as_current_span("foo"):
+            with tracer.start_as_current_span("bar"):
+                with tracer.start_as_current_span("baz"):
                     # add span processor after spans have been created
                     tracer.add_span_processor(sp)