+++ opentelemetry-sdk/CHANGELOG.md
@@ -1,4 +1,21 @@
 # Changelog
 
 ## Unreleased
+
+## 0.2a0
+
+Released 2019-10-29
+
+- W3C TraceContext fixes and compliance tests
+  ([#228](https://github.com/open-telemetry/opentelemetry-python/pull/228))
+- Multiple metrics SDK changes
+- Multiple tracing SDK changes
+- Sampler SDK
+  ([#225](https://github.com/open-telemetry/opentelemetry-python/pull/225))
+- Multiple bugfixes and improvements
+
+## 0.1a0
+
+Released 2019-09-30
+
 - Initial release